feat(backend): enforce strict security checks and hardened admin access controls
- Add strict validation for CCXT exchange client operations
- Implement admin middleware security hardening
- Enhance error recovery with better validation
- Strengthen notification service with input sanitization
- Add test coverage for notification security fixes
- Refine signal processor with defensive programming

All changes improve security posture across backend services
to prevent injection vulnerabilities and unauthorized access.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-opencode)

+++ services/backend-api/internal/services/signal_processor.go
@@ -300,7 +300,7 @@ func (sp *SignalProcessor) processSignalBatchWithRetry(ctx context.Context, star
 			}).Info("Retrying signal batch processing")
 
 			// Wait before retry with exponential backoff
-			retryDelay := sp.config.RetryDelay * time.Duration(1<<uint(attempt-1))
+			retryDelay := sp.config.RetryDelay * time.Duration(1<<(attempt-1))
 			select {
 			case <-ctx.Done():
 				return ctx.Err()
